Expands mypy coverage from 14 to 29 source files:
- Source-types get_season_progress with SeasonProgress TypedDict
- Adds clean modules (sgp/*, analysis/*, utils/*) to [tool.mypy].files
- Annotates simulation.py locals, utils/positions.py None guard,
  analysis/pace.py float cast, lineup/matchups.py cast + result,
  lineup/optimizer.py best tuple + partial_best rename

+++ src/fantasy_baseball/analysis/pace.py
@@ -113,7 +113,7 @@ def compute_player_pace(
         dev = (rest_of_season_val - pre_val) / denom
         if cat in INVERSE_STATS:
             dev = -dev
-        return round(dev, 2)
+        return float(round(dev, 2))
 
     # Opportunity column (PA or IP) — always neutral
     result[opp_key.upper()] = {
